requirements.txt
mirascope==1.2.0
mirascope[groq]==1.2.0
mirascope[gemini]==1.2.0
mirascope[openai]==1.2.0
mirascope[tenacity]==1.2.0
tenacity
qdrant-client==1.11.1
python-dotenv==1.0.1
lancedb==0.12.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-groq
langchain-ollama
numpy
ollama==0.3.2
pydantic==2.7.4
fastembed==0.3.6
groq==0.11.0